Several improvements.
Site is determined from the individual URI.
Site may be specified or not. If specified, it restricts the counts and the locate queries.

Add a RESTART option that will clear the Solr index and the bookmark.

Add counts to the report.

Remove the commit after each call to SolrServer::add_document.